chore: move auth route definition to auth module
