+++ app-modules/auth/src/Http/Requests/ProfileUpdateRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace Modules\Auth\Http\Requests;
 
 use App\Models\User;
 use Illuminate\Foundation\Http\FormRequest;
